pyproject.toml
[tool.poetry]
name = "slices-clincal-trial"
version = "0.1.0"
description = "Python Backend"
authors = ["Saad Shirgaonkar <saad.shirgaonkar@siroclinpharm.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
motor = "^3.6.0"
uvicorn = "^0.34.0"
bs4 = "^0.0.2"
openai = "^1.59.6"
numpy = "^2.2.1"
pinecone-client = "^5.0.1"
transformers = "^4.48.0"
scikit-learn = "^1.6.1"
python-dotenv = "^1.0.1"
pytz = "^2025.1"
boto3 = "^1.37.7"
pypdf2 = "^3.0.1"
python-multipart = "^0.0.20"
pytest = "^8.3.5"
redis = "^5.2.1"
rq = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
